First Commit
The College Book Space was an idea I had last year(2011). The Idea was
to create a website where students could sell their textbooks locally. A
sort of craigslist for textbooks. You can see the home page at
CollegeBookSpace.com. The site is not fully functional but its skelaton
is there. I learned quite a bit while working on this project.